Restructure test module - extract all tests to separate modules
